pyproject.toml
[project]
name = "simple-flotilla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "numpy>=2.3.5",
]
